# lib/api-zod/src/generated/types/attendee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Event Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Attendee {
  id: number;
  eventId: number;
  /** @nullable */
  displayName?: string | null;
  assignedId: number;
  joinedAt: Date;
  raisedHand: boolean;
  /** @nullable */
  raisedHandAt?: Date | null;
}