# lib/api-zod/src/generated/types/attendeeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Event Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Attendee } from "./attendee";

export interface AttendeeResponse {
  attendee: Attendee;
}