# lib/api-zod/src/generated/types/createEventBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Event Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEventBody {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  promoText?: string | null;
  /** @nullable */
  startTime?: Date | null;
}