# lib/api-zod/src/generated/types/hostStatsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Event Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface HostStatsResponse {
  totalEvents: number;
  liveEvents: number;
  totalAttendees: number;
}