# lib/api-zod/src/generated/types/joinEventBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Event Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface JoinEventBody {
  /** @nullable */
  displayName?: string | null;
}