# lib/api-zod/src/generated/types/joinEventResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Live Event Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Attendee } from "./attendee";
import type { Event } from "./event";

export interface JoinEventResponse {
  attendee: Attendee;
  event: Event;
}